Structured a bit
